+++ extensions/mute-microphone/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Mute microphone Changelog
 
+## [Windows Support] - 2026-01-19
+
+- Added support for Windows using PowerShell scripts
+- Implemented platform detection to switch between macOS (AppleScript) and Windows (PowerShell)
+- Added 'Set Level' support for Windows
+
 ## [Bugfix] - 2024-10-16
 
 - Fix overlapping toast messages